Extends another test for better branch coverage
