+++ src/onegov/activity/models/invoice.py
@@ -5,12 +5,14 @@ from onegov.pay import Invoice
 from onegov.user import User
 from sqlalchemy import Column
 from sqlalchemy import ForeignKey
+from sqlalchemy.ext.hybrid import hybrid_property
 from sqlalchemy.orm import relationship
 
 
 from typing import Any, TYPE_CHECKING
 if TYPE_CHECKING:
     import uuid
+    from collections.abc import Iterable
     from decimal import Decimal
     from onegov.activity.models import ActivityInvoiceItem
     from onegov.activity.models import BookingPeriod
@@ -47,6 +49,13 @@ class BookingPeriodInvoice(Invoice):
     if TYPE_CHECKING:
         items: relationship[list[ActivityInvoiceItem]]  # type: ignore[assignment]
 
+    @property
+    def has_donation(self) -> bool:
+        for item in self.items:
+            if item.group == 'donation':
+                return True
+        return False
+
     def add(
         self,
         group: str,
@@ -70,3 +79,53 @@ class BookingPeriodInvoice(Invoice):
             flush=flush,
             **kwargs
         )
+
+    @hybrid_property
+    def discourage_changes(self) -> bool:
+        return self.discourage_changes_for_items(self.items)
+
+    @hybrid_property
+    def disable_changes(self) -> bool:
+        return self.disable_changes_for_items(self.items)
+
+    @hybrid_property
+    def has_online_payments(self) -> bool:
+        return self.has_online_payments_for_items(self.items)
+
+    def discourage_changes_for_items(
+        self,
+        items: Iterable[ActivityInvoiceItem]
+    ) -> bool:
+        for item in items:
+            if item.source == 'xml':
+                return True
+
+        return False
+
+    def disable_changes_for_items(
+        self,
+        items: Iterable[ActivityInvoiceItem]
+    ) -> bool:
+        for item in items:
+            if not item.source:
+                continue
+
+            if item.source == 'xml':
+                continue
+
+            states = {p.state for p in item.payments}
+
+            if 'open' in states or 'paid' in states:
+                return True
+        return False
+
+    def has_online_payments_for_items(
+        self,
+        items: Iterable[ActivityInvoiceItem]
+    ) -> bool:
+        for item in items:
+            if not item.source or item.source == 'xml':
+                continue
+
+            return True
+        return False
